Rename to dbg_interfaces.rs for clarity
